feat: Add a script to generate deprecated attributes (#61)
Adds a script to generate a JSON of deprecated fields
so the ingest pipeline can use sentry-conventions
as the source of truth.

Also adds a _status field to the attributes schema.
We can use this internally do decide which stage of the
deprecation an is in. Deprecated fields need a 90 day
backfill (ie relay double-writes, EAP coalesces etc) after
which they can just be normalized to the new field.