test(name): verify attribute usage in template lists (#153)
To assist the code generation in Relay, it is best if name template lists
always contain a single entry without attributes as a fallback, and that that
entry is the final template in the list.

This is already the case for our existing templates, but this test enforces it
going forward as well.