ref: Undeprecate `sentry.observed_timestamp_nanos` (#174)
https://github.com/getsentry/sentry-conventions/pull/137 added
`sentry.observed_timestamp_nanos` as a deprecated attribute and
`sentry._internal.observed_timestamp_nanos` as its non-deprecated
replacement. However, adding the attribute as deprecated was probably a
mistake because it is in active use and no concrete plan was made for a
migration to the `_internal` version. It's also questionable whether
this attribute should be hidden from users in the first place.

Since the `_internal` version is not used anywhere and not written by
Relay, I propose that we delete it and mark the non-`_internal` version
as not deprecated.